kernel: AF_UNIX matches Linux + arm64 busybox init bring-up
Three closely-related bug fixes that together let real Alpine
production busybox boot and run as init on arm64:

1. clone(2): when CLONE_SETTLS is unset, capture the live
   TPIDR_EL0 register so children inherit the parent's TLS
   pointer (matches the clone(2) man page).  Without this,
   posix_spawn'd children landed with TPIDR_EL0=0 and NULL-derefed
   on the first errno write inside musl's __syscall_ret.

2. SIGMAX widened 32 → 64 (full Linux NSIG).  musl's __init_libc
   calls rt_sigaction on RT signals during startup; the old
   SIGMAX returned EINVAL and triggered the chain crash above.
   Cascading u32 → u64 across signal_pending, signalfd mask,
   rt_sigtimedwait mask, procfs SigPend reporting, and the
   trait method on Platform.

3. AF_UNIX bind(2) now creates a real S_IFSOCK inode at the path
   (via tmpfs/ext2 create_file honoring the mode type bits) AND
   the listener registry is keyed by (dev_id, inode_no, generation)
   instead of by path string.  Effects:

     - `[ -S /tmp/.X11-unix/X0 ]` succeeds; stat() reports S_IFSOCK.
     - readdir returns DT_SOCK (added FileType::Socket = 12).
     - mv of a bound socket follows the inode; connect to the new
       path routes correctly.
     - unlink + connect → ENOENT (Linux's exact errno).
     - bind to a stale path → EADDRINUSE.
     - Generation counter disambiguates ext2 inode-number reuse
       so a long-lived listener whose path was unlinked can't
       collide with a different file later bound at the same ino.

Also removed the per-process cached_utsname snapshot: each fork
copied it from the parent at fork time, so a sethostname() in
one process didn't propagate to siblings sharing the same UTS
namespace.  Linux always reads from the live namespace; we now
do too.  uname() rebuilds the 390-byte struct on each call from
the (shared) Arc<UtsNamespace>.

Smaller fixes folded in:
- page_fault.rs: dump R8/R9 (= aarch64 x7/x8, the syscall number
  is in x8) alongside other GPRs in fatal SIGSEGV.  This is what
  pointed at the rt_sigaction → __syscall_ret chain in (2) above.

Verified on arm64:
- test-lxde 6/6 (full LXDE session under Kevlar)
- test-threads-smp 14/14
- test-busybox-alpine 106/106 (production Alpine busybox; same
  result as Linux, byte-identical)

+++ kernel/syscalls/rt_sigtimedwait.rs
@@ -43,7 +43,7 @@ impl<'a> SyscallHandler<'a> {
             b[..4].copy_from_slice(&raw);
             b
         };
-        let mask = u64::from_le_bytes(set_bytes) as u32;
+        let mask = u64::from_le_bytes(set_bytes);
 
         // Parse timeout.
         let deadline_ns: Option<u64> = if let Some(tp) = timeout_ptr {
@@ -121,7 +121,7 @@ impl<'a> SyscallHandler<'a> {
 }
 
 /// Try to dequeue a signal matching `mask` from the current process.
-fn try_dequeue_signal(mask: u32) -> Option<Signal> {
+fn try_dequeue_signal(mask: u64) -> Option<Signal> {
     let current = current_process();
     let mut signals = current.signals().lock();
     signals.pop_pending_masked(mask)
